chore(stripe): Update OpenAPI
