+++ .gitattributes
@@ -1,0 +1,2 @@
+# OpenAPI files that live under scripts/openapi/<name>/...
+scripts/openapi/stripe/internal/files/spec3.yaml filter=lfs diff=lfs merge=lfs -text
