--- providers/stripe/openapi/README.md
@@ -1,7 +1,0 @@
-# OpenAPI
-
-## Obtain file
-
-You can download the file from official GitHub repo https://github.com/stripe/openapi.
-
-The yaml file uses OpenAPI version 3.
