--- providers/stripe/openapi/file.go
@@ -1,16 +1,0 @@
-package openapi
-
-import (
-	_ "embed"
-
-	"github.com/amp-labs/connectors/tools/fileconv/api3"
-)
-
-var (
-	// Static file containing openapi spec.
-	//
-	//go:embed spec3.yaml
-	apiFile []byte
-
-	FileManager = api3.NewOpenapiFileManager[any](apiFile) // nolint:gochecknoglobals
-)
