+++ scripts/openapi/stripe/internal/files/README.md
@@ -1,0 +1,8 @@
+# OpenAPI
+
+## Obtain file
+
+You can download the file from official GitHub repo https://github.com/stripe/openapi.
+For example: https://github.com/stripe/openapi/blob/master/openapi/spec3.yaml
+
+The yaml file uses OpenAPI version 3.
