+++ scripts/openapi/stripe/internal/files/spec3.yaml
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:aa68d0aba2ced32350bdb71389ebce6c6dd885bdda1505a39ff175ba008865c7
+size 6281081
